Quad Tree V1
New Class XDL_QuadTree, I coded a quick quad tree implementation, with
debug lines to show its various regions.

Going to be used for objects added to scene. Quad tree currently seems
to be working when I add Objects to the scene. Going to leverage
quadtree to enable draw culling, and possible update culling

By default only objects onscreen will be drawn, but all objects will be
updated. Hoping to give user option to tag objects to be not updated
after the go offscreen. This will be usefull for Animation managers etc,
and things that don't need to be happening untill they are onscreen.

Quadtree is in, but it is not doing anything but sorting objects into
children in its tree.